Address review: unify memory_limit type to u32, simplify estimation
- Change DEFAULT_MEMORY_LIMIT_MB, PrefetchConfig.memory_limit_mb, and
  MemoryBudget API from u64 to u32, eliminating the lossy downcast
- Inline the 1.5x multiplier with a doc comment instead of named constants
- Regenerate connection_config.py after rebase onto main

+++ sf_core/src/apis/database_driver_v1/statement.rs
@@ -831,7 +831,7 @@ async fn resolve_prefetch_config(conn: &Arc<Mutex<Connection>>) -> PrefetchConfi
         .unwrap_or(crate::chunks::DEFAULT_PREFETCH_THREADS);
     let memory_limit_mb = session_params
         .get(param_names::CLIENT_MEMORY_LIMIT.as_str())
-        .and_then(|v| v.parse::<u64>().ok())
+        .and_then(|v| v.parse::<u32>().ok())
         .unwrap_or(crate::chunks::DEFAULT_MEMORY_LIMIT_MB);
     PrefetchConfig {
         prefetch_threads: threads,
